backend/requirements.txt
alembic==1.13.2
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
google-auth==2.33.0
google-auth-oauthlib==1.2.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
Mako==1.3.5
MarkupSafe==2.1.5
msgpack==1.0.8
numpy==2.0.0
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
SQLAlchemy==2.0.32
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
